Added in password encryption and decryption

+++ BugTrackerCommandLine/Users.cs
@@ -10,22 +10,28 @@ namespace BugTrackerCommandLine
             Console.WriteLine("-----------------------------------------------");
             Console.WriteLine("List Users");
 
-            var sql = "select user_name, display_name, email, is_active from Users;";
+            var sql = "select user_name, display_name, email, password, is_active from Users;";
 
             DataEntry data = new DataEntry();
             data.ConnectToDatabase();
             var results = data.RunSQL(sql);
 
-            foreach (DataRow row in results.Rows)
+            if (results.Rows.Count > 0)
             {
-                Console.WriteLine("-----------------------------------------------");
-                Console.WriteLine("User Name: " + row.Field<string>("user_name"));
-                Console.WriteLine("Display Name: " + row.Field<string>("display_name"));
-                Console.WriteLine("Email: " + row.Field<string>("email"));
-                Console.WriteLine("Is Active: " + row.Field<bool>("is_active"));
-                Console.WriteLine("-----------------------------------------------");
-                Console.WriteLine("");
+                foreach (DataRow row in results.Rows)
+                {
+                    Console.WriteLine("-----------------------------------------------");
+                    Console.WriteLine("User Name: " + row.Field<string>("user_name"));
+                    Console.WriteLine("Display Name: " + row.Field<string>("display_name"));
+                    Console.WriteLine("Email: " + row.Field<string>("email"));
+                    Console.WriteLine("Password: " + row.Field<string>("password"));
+                    Console.WriteLine("Is Active: " + row.Field<bool>("is_active"));
+                    Console.WriteLine("-----------------------------------------------");
+                    Console.WriteLine("");
+                }
             }
+            else
+                Console.WriteLine("No Users Found");
 
             Console.WriteLine("Press Any Key to Continue");
             Console.ReadKey();
@@ -53,9 +59,11 @@ namespace BugTrackerCommandLine
             Console.WriteLine("Enter Password:");
             password = Console.ReadLine();
 
-            string sql = $"INSERT INTO Users (user_name, display_name, email, password, is_active) VALUES  ('{name}', '{display_name}', '{email}', '{password}', 1)";
-
             DataEntry data = new DataEntry();
+            string encryptedPw = EncryptionDecryptionService.Encrypt(data.getResource("key"), password);
+
+            string sql = $"INSERT INTO Users (user_name, display_name, email, password, is_active) VALUES  ('{name}', '{display_name}', '{email}', '{encryptedPw}', 1)";
+
             data.ConnectToDatabase();
             data.RunSQL(sql);
 
@@ -74,16 +82,21 @@ namespace BugTrackerCommandLine
             data.ConnectToDatabase();
             var results = data.RunSQL(sql);
 
-            foreach (DataRow row in results.Rows)
+            if (results.Rows.Count > 0)
             {
-                Console.WriteLine("-----------------------------------------------");
-                Console.WriteLine("ID: " + row.Field<int>("id"));
-                Console.WriteLine("User Name: " + row.Field<string>("user_name"));
-                Console.WriteLine("Display Name: " + row.Field<string>("display_name"));
-                Console.WriteLine("Email: " + row.Field<string>("email"));
-                Console.WriteLine("-----------------------------------------------");
-                Console.WriteLine("");
+                foreach (DataRow row in results.Rows)
+                {
+                    Console.WriteLine("-----------------------------------------------");
+                    Console.WriteLine("ID: " + row.Field<int>("id"));
+                    Console.WriteLine("User Name: " + row.Field<string>("user_name"));
+                    Console.WriteLine("Display Name: " + row.Field<string>("display_name"));
+                    Console.WriteLine("Email: " + row.Field<string>("email"));
+                    Console.WriteLine("-----------------------------------------------");
+                    Console.WriteLine("");
+                }
             }
+            else
+                Console.WriteLine("No Users Found");
 
             Console.WriteLine("Enter the ID you want to modify:");
             string id = Console.ReadLine();
@@ -102,13 +115,15 @@ namespace BugTrackerCommandLine
             Console.WriteLine($"Enter new Password for ID {id}:");
             password = Console.ReadLine();
 
+            string encryptedPw = EncryptionDecryptionService.Encrypt(data.getResource("key"), password);
+
             Console.WriteLine($"Is {id} still Active? (y/n)");
             if (Console.ReadLine().ToLower() == "y")
                 is_active = true;
             else
                 is_active = false;
 
-            sql = $"update Users set user_name='{name}', display_name='{display_name}', email='{email}', password='{password}', is_active={is_active}";
+            sql = $"update Users set user_name='{name}', display_name='{display_name}', email='{email}', password='{encryptedPw}', is_active={is_active}";
             data.ConnectToDatabase();
             data.RunSQL(sql);
 
@@ -127,28 +142,36 @@ namespace BugTrackerCommandLine
             data.ConnectToDatabase();
             var results = data.RunSQL(sql);
 
-            foreach (DataRow row in results.Rows)
+            if (results.Rows.Count > 0)
             {
-                Console.WriteLine("-----------------------------------------------");
-                Console.WriteLine("ID: " + row.Field<int>("id"));
-                Console.WriteLine("User Name: " + row.Field<string>("user_name"));
-                Console.WriteLine("Display Name: " + row.Field<string>("display_name"));
-                Console.WriteLine("-----------------------------------------------");
-                Console.WriteLine("");
+                foreach (DataRow row in results.Rows)
+                {
+                    Console.WriteLine("-----------------------------------------------");
+                    Console.WriteLine("ID: " + row.Field<int>("id"));
+                    Console.WriteLine("User Name: " + row.Field<string>("user_name"));
+                    Console.WriteLine("Display Name: " + row.Field<string>("display_name"));
+                    Console.WriteLine("-----------------------------------------------");
+                    Console.WriteLine("");
+                }
+
+                Console.WriteLine("Enter the ID you want to delete:");
+                string id = Console.ReadLine();
+
+                Console.WriteLine($"Are you sure you would like to delete ID {id}? (y/n)");
+                string answer = Console.ReadLine();
+
+                if (answer.ToLower() == "y")
+                {
+                    sql = $"DELETE FROM Users WHERE id = {id}";
+                    data.ConnectToDatabase();
+                    data.RunSQL(sql);
+                }
             }
-
-            Console.WriteLine("Enter the ID you want to delete:");
-            string id = Console.ReadLine();
-
-            Console.WriteLine($"Are you sure you would like to delete ID {id}? (y/n)");
-            string answer = Console.ReadLine();
-
-            if(answer.ToLower() == "y")
+            else
             {
-                sql = $"DELETE FROM Users WHERE id = {id}";
-                data.ConnectToDatabase();
-                data.RunSQL(sql);
-            }                 
+                Console.WriteLine("No Users Found");
+                Console.ReadKey();
+            }
 
             Console.WriteLine("-----------------------------------------------");
             Console.WriteLine("");
